requirements.txt
astroid==2.1.0
autopep8==1.4.3
boto3==1.9.91
colorama==0.4.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.1.5
django-heroku==0.3.1
django-storages==1.7.1
gunicorn==19.9.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.4.1
psycopg2==2.7.7
pycodestyle==2.5.0
pylint==2.2.2
pytz==2018.9
six==1.12.0
static3==0.7.0
whitenoise==4.1.2
wrapt==1.11.1
djangorestframework==3.9.1
django-cors-headers==2.4.0
